add dynamic umd_glad_sentinel2_alerts, update dependencies
