+++ Pipfile
@@ -20,7 +20,7 @@ mercantile = "==1.2.1" # same as lambda layer
 aiofiles = "*"
 aenum = "*"
 asyncpg = "*"
-boto3 = "==1.16.52"  # aioboto3 dependency
+boto3 = "===1.17.49"  # aioboto3 dependency
 fastapi = "*"
 mercantile = "*"
 uvicorn = "*"
